Add files via upload
Question:
Write a program in C++ to implement a stack using an array. Your program should provide the following operations:

Push: Add a new element to the top of the stack. Display an error message if the stack is full.
Pop: Remove and display the top element of the stack. Display an error message if the stack is empty.
Show: Display all the elements currently in the stack from top to bottom.